backend/requirements.txt
Django==3.1.2
djangorestframework==3.12.2
requests==2.24.0